Extract replay probe rotation analysis
